Added Different Sprites to Get Different Points
I've added some sprites to make differents types of scoring, positive and negative. Also I've add the functionality to difference when the player crash into a good point item and a bad point item.  Also I fixed a problem in spawn manager that was starting a coroutine from the beginning, without waiting seconds.